requirements.txt
psycopg2~=2.9.3
loguru~=0.6.0
uvicorn~=0.17.6
fastapi~=0.75.1
sqlalchemy~=1.4.35
python-jose[cryptography]